requirements.txt
python-telegram-bot==21.5
anthropic==0.34.2
google-genai==0.6.0
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
aiofiles==24.1.0
gtts==2.5.3
python-dotenv==1.0.1
httpx==0.27.2
